feat: 添加敌人路径跟随及随机工具函数
- 在 math.hpp 中添加 `range_random` 工具函数，用于线程安全的随机数生成
- 在 CMakeLists.txt 中包含新的游戏系统源文件（followpath_system, remove_dead_system）
- 更新移动系统，直接使用 `delta.asSeconds()` 而非静态转换
- 为 GameScene 扩展敌人路径跟随功能：
  - 添加 FollowPathSystem 和 RemoveDeadSystem
  - 实现敌人到达事件的事件连接
  - 添加用于敌人导航的路径点节点管理
  - 包含测试敌人的创建方法
- 为游戏系统和事件添加必要的头文件包含及前向声明

+++ src/engine/system/movement_system.cpp
@@ -12,7 +12,7 @@ void MovementSystem::update(entt::registry& registry, sf::Time delta) {
 
     // 遍历获取的实体，获取组件并执行相关逻辑
     for (auto&& [entity, velocity, transform] : view.each()) {
-        transform.position_ += velocity.velocity_ * static_cast<float>(delta.asSeconds());
+        transform.position_ += velocity.velocity_ * delta.asSeconds();
     }
 }
